+++ shoop/admin/modules/service_providers/views/__init__.py
@@ -6,10 +6,12 @@
 # This source code is licensed under the AGPLv3 license found in the
 # LICENSE file in the root directory of this source tree.
 
+from ._delete import ServiceProviderDeleteView
 from ._edit import ServiceProviderEditView
 from ._list import ServiceProviderListView
 
 __all__ = [
+    "ServiceProviderDeleteView",
     "ServiceProviderEditView",
     "ServiceProviderListView",
 ]